+++ DentistApp.UI/ViewModels/OperationViewModel.cs
@@ -72,6 +72,7 @@ namespace DentistApp.UI.ViewModels
         }
         public void DeleteOperation(object o)
         {
+            if (ShouldDelete()) return;
             var operation = o as Operation;
             OperationController.Delete(Operation);
             LoadOperations(null);
